+++ karma.conf.js
@@ -9,8 +9,12 @@ var webpackConfig = {
   module: {
     preLoaders: [{
       test: /.*\.js$/,
-      loaders: ['eslint'],
+      loader: 'eslint',
       exclude: /node_modules/
+    }],
+    loaders: [{
+      test: /\.html$/,
+      loader: 'raw'
     }],
     postLoaders: [{
       test: /\.js$/,