+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.6.0 (2015-11-05)
+* Allow the default popover template to be changed either globally in the `confirmationPopoverDefaults.templateUrl` property or by passing a `template-url` property to the popover.
+* Expose the popover element on the controller so other directives can access it
+
 # 0.5.2 (2015-10-24)
 * Don't leak the module name characters into the global space